created utils folder and moved files
