+++ src/pages/questionPage.js
@@ -8,7 +8,7 @@ import { createAnswerComponent } from '../components/answerComponent.js';
 import { questionNumberTracker } from '../components/questionNumTracker.js';
 import { quizData } from '../data.js';
 import { initResultPage } from './resultPage.js';
-import { setLocalStorage } from '../components/setLocalStorage.js';
+import { setLocalStorage } from '../utils/setLocalStorage.js';
 
 export const initQuestionPage = () => {
   const userInterface = document.getElementById(USER_INTERFACE_ID);
